Allow deletion of scheduled events
Scheduled events can be deleted for usability purposes. In order for an
event to be deleted, it needed to be assigned an event handler, but the
event handler needed a key so it can be identified by the user. This
created the need for an events queue dictionary to track this
information.

An events queue needs to be in sync with the scheduler's queue. This
created several situations where the two can be out-of-sync. In the
future, an automated testing solution will be needed in order to keep
track of these types of issues.

- Create events_queue and unique IDs assigned to each scheduled event

- Added a few commands such as 'delete' and 'quit'

- Created a class to keep track of event details